+++ include/fastrtps/rtps/timedevent/DeadlineTimer.h
@@ -21,7 +21,7 @@
 #define DeadlineTimer_H_
 #ifndef DOXYGEN_SHOULD_SKIP_THIS_PUBLIC
 
-#include "TimedEvent.h"
+#include "../resources/TimedEvent.h"
 
 #include <functional>
