+++ src/cpp/CMakeLists.txt
@@ -31,7 +31,7 @@ set(${PROJECT_NAME}_source_files
     rtps/resources/TimedEventImpl.cpp
     rtps/resources/AsyncWriterThread.cpp
     rtps/resources/AsyncInterestTree.cpp
-    rtps/resources/DeadlineTimer.cpp
+    rtps/timedevent/DeadlineTimer.cpp
     rtps/writer/RTPSWriter.cpp
     rtps/writer/StatefulWriter.cpp
     rtps/writer/ReaderProxy.cpp
